+++ Changelog.md
@@ -7,6 +7,9 @@ The next release will include the following feature enhancements and bug fixes:
 **Features**
 - Added Default Connection support (https://github.com/aws/graph-explorer/pull/108)
 
+**Bug fixes**
+- Fixed Gremlin/openCypher matching ID property on all keyword searches (https://github.com/aws/graph-explorer/pull/169)
+
 ## Release 1.3.1
 
 This patch release includes bugfixes for Release 1.3.0.